Move common range_circles into module
